+++ Task_2_1_2/src/main/java/ru/nsu/sartakov/employee/Deliverer.java
@@ -7,6 +7,10 @@ import ru.nsu.sartakov.queue.SharedQueue;
 
 import java.util.List;
 
+/**
+ * The Deliverer class simulates the work of deliverer
+ * Get the order from shared storage and deliver it to abstract customer
+ */
 public class Deliverer implements Runnable, Consumer<List<Order>> {
     private final int deliveryTime = 300;
     private final int id;
@@ -15,16 +19,29 @@ public class Deliverer implements Runnable, Consumer<List<Order>> {
     private final SharedQueue<Order> storage;
     private boolean isRunning;
 
+    /**
+     * Constructor for class Deliverer
+     * @param id - deliverer's id
+     * @param capacity - capacity of Deliverer (how many order can it take at once)
+     * @param storage - shared order storage
+     */
     public Deliverer(int id, int capacity, SharedQueue<Order> storage){
         this.id = id;
         this.capacity = capacity;
         this.storage = storage;
     }
 
+    /**
+     * Stop the working process
+     */
     public void stop() {
         this.isRunning = false;
     }
 
+    /**
+     * Method returning the id of deliverer
+     * @return deliverer id
+     */
     public int getId() {
         return this.id;
     }
@@ -35,6 +52,11 @@ public class Deliverer implements Runnable, Consumer<List<Order>> {
         }
     }
 
+    /**
+     * Get finished order from shared storage
+     * Reduces the shared storage
+     * @return delivered orders
+     */
     @Override
     public List<Order> consume() {
         try {
@@ -48,18 +70,19 @@ public class Deliverer implements Runnable, Consumer<List<Order>> {
         }
     }
 
-    public void work() {
-        List<Order> orders = consume();
-        if (orders == null) {
-            stop();
-        }
-    }
-
+    /**
+     * Start a thread of deliverer
+     * Method simulates working process
+     * Till the shared storage is not empty, Deliverer will work (running)
+     */
     @Override
     public void run() {
         this.isRunning = true;
         while(this.isRunning) {
-            work();
+            List<Order> orders = consume();
+            if (orders == null) {
+                stop();
+            }
         }
     }
 }
