--- Task_2_1_2/src/main/java/ru/nsu/sartakov/json/BakerJSON.java
@@ -1,14 +1,0 @@
-package ru.nsu.sartakov.json;
-
-public class BakerJSON {
-    private int id;
-    private int cookingTime;
-
-    public int getId() {
-        return id;
-    }
-
-    public int getCookingTime() {
-        return cookingTime;
-    }
-}
