--- Task_2_1_2/src/main/java/ru/nsu/sartakov/json/DelivererJSON.java
@@ -1,14 +1,0 @@
-package ru.nsu.sartakov.json;
-
-public class DelivererJSON {
-    private int id;
-    private int capacity;
-
-    public int getId() {
-        return id;
-    }
-
-    public int getCapacity() {
-        return capacity;
-    }
-}
