+++ Task_2_1_2/src/main/java/ru/nsu/sartakov/json/Json.java
@@ -5,26 +5,45 @@ import com.google.gson.GsonBuilder;
 
 import java.io.*;
 
+/**
+ * Class for reading JSON file
+ */
 public class Json {
     private static final String DEFAULT = "pizzeriaConfig/config.json";
     private final File file;
     private final String fileName;
     private BufferedReader reader;
 
+    /**
+     * Constructor for Json class
+     * Initialize the file with DEFAULT file name
+     */
     public Json() {
         this.fileName = DEFAULT;
         this.file = new File(this.fileName);
     }
 
+    /**
+     * Constructor for Json class
+     * Initialize the file with specific fie name
+     * @param fileName - Name of config file
+     */
     public Json(String fileName) {
         this.fileName = fileName;
         this.file = new File(fileName);
     }
 
+    /**
+     * @return file name
+     */
     public String getFileName() {
         return fileName;
     }
 
+    /**
+     * Opens the file, make new file if not exist
+     * Opens the file for reading
+     */
     public void open() {
         try {
             if (!this.file.exists()) {
@@ -48,6 +67,10 @@ public class Json {
         return content.toString();
     }
 
+    /**
+     * Reads the pizzeria config from JSON
+     * @return the array of parsable configuration settings
+     */
     public PizzeriaJSON read() {
         String content;
         try {
@@ -64,6 +87,9 @@ public class Json {
         return gson.fromJson(content, PizzeriaJSON.class);
     }
 
+    /**
+     * Closes the file buffered reader
+     */
     public void close() {
         try {
             reader.close();
